requirements.txt
fastapi
uvicorn
requests
PyJWT 
PyGithub
starlette
itsdangerous